Split up common and web part of certificate controller
